#24 #23 Moved Nav to separate component and updated details page styling
