+++ src/frontend/Details/Details.css
@@ -1,7 +1,10 @@
 
 body {
   background-color: #A1D6E2;
-  font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
+}
+
+.details-title {
+  color: black;
 }
 
 .details-container {
