+++ src/frontend/header/header.js
@@ -1,32 +1,13 @@
 import React from 'react';
 import ReactTextRotator from 'react-text-rotator';
 import './header.css';
-import CryptoCard from '../crypto/crypto';
+import Nav from '../Nav/Nav';
 
 
 export default function Header() {
-  const coinData = {
-    name: 'Bitcoin',
-    symbol: 'BTC',
-  }
-
     return (
-      <div className="Header" id="header-section">
-        <div className="nav-container">
-          <div className="container">
-            <div className="nav-bar">
-              <div className="header-logo">
-                <header>8 Days a Week</header>
-              </div>
-              <nav id="navigation">
-                <CryptoCard
-                  name={coinData.name}
-                  symbol={coinData.symbol}
-                />
-              </nav>
-            </div>
-          </div>
-        </div>
+      <div className='Header'>
+        <Nav />
         <div className="container">
           <div className="welcome-area">
             <div className="header-text text-center">
